Backend/requirements.txt
bcrypt
pyjwt
pymongo
Flask
flask_cors